==> jupyter_notebooks/Modelo 11/02_gestion_nulos.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import numpy as np
import pandas as pd
pd.set_option('display.float_format', '{:.2f}'.format)


# Imputación de nulos usando métodos avanzados estadísticos
# -----------------------------------------------------------------------
from sklearn.impute import SimpleImputer
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn.impute import KNNImputer
from sklearn.ensemble import RandomForestRegressor

import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns',100)


df = pd.read_csv("../../datos/02_api_rent_limpieza_cols.csv")
df.sample()

# Gestión de Valores erróneos
Vamos a revisarlos en:
- Precio
- pricebyarea
- size
- rooms
- bathrooms


df["province"].unique()

### Modelo 7 Eliminar casas ajenas a "Madrid" 

df = df[df["province"] == "Madrid"]
df.reset_index(drop=True,inplace=True)
df.shape

df["price"].unique()

No se ven anómalos

### size

df["size"].nlargest(5)

df["size"].nsmallest(5)

No parece erróneos

### rooms

df["rooms"].unique()

0 es que son estudios

### bathrooms

df["bathrooms"].unique()

No se ven valores erróneos

# Corregir "floor"
- Vamos a pasar todo a texto para que sea una categórica

df["floor"].unique()

### Generamos un diccionario para aplicar los cambios

diccionario_pisos = {
    "1": "primero",
    "2": "segundo",
    "3": "tercero",
    "4": "cuarto",
    "5": "quinto",
    "6": "sexto",
    "7": "septimo",
    "8": "octavo",
    "14": "decimo cuarto",
    "bj": "bajo",
    "en": "entreplanta",
    "ss": "sotano",
    "st": "sotano",
}

df["floor"] = df["floor"].map(diccionario_pisos)
df["floor"].unique()


# Corregir Bathrooms
- Vamos a pasar todo a texto para que sea una categórica

df["bathrooms"].unique()

df.loc[df["bathrooms"] == 1, "bathrooms"] = "1 aseo"
df.loc[df["bathrooms"] == 2, "bathrooms"] = "2 aseos"
df.loc[df["bathrooms"] == 3, "bathrooms"] = "3 aseos"

df["bathrooms"].unique()

# Corregir rooms
- Vamos a pasar todo a texto para que sea una categórica

df["rooms"].unique()

df.loc[df["rooms"] == 0, "rooms"] = "sin habitaciones"
df.loc[df["rooms"] == 1, "rooms"] = "1 habitacion"
df.loc[df["rooms"] == 2, "rooms"] = "2 habitaciones"
df.loc[df["rooms"] == 3, "rooms"] = "3 habitaciones"
df.loc[df["rooms"] == 4, "rooms"] = "4 habitaciones"
df.loc[df["rooms"] == 6, "rooms"] = "6 habitaciones"

df["rooms"].unique()

# Reducción / Simplificación de columnas
- Vamos a quitarnos todo lo que:
    - Especifique demasiado para el precio
- Vamos a simplificar todo lo que:
    - Especifique demasiado, pero al generalizarlo, aporte valor

df.sample()

### Dropeamos
- address: No se puede generalizar

df.drop(columns="address",inplace=True)

### Generalizamos
- distance: Lo haremos por rangos de distancia al centro

df["distance"].nlargest()

df["distance"].nsmallest()

df["distancia_centro"] = "unknown"
df.loc[df["distance"].between(0,1000,inclusive="left"), "distancia_centro"] = "Menos de 1 km"
df.loc[df["distance"].between(1000,5000,inclusive="left"), "distancia_centro"] = "Entre 1 y 5 km"
df.loc[df["distance"].between(5000,10000,inclusive="left"), "distancia_centro"] =  "Entre 5 y 10 km"
df.loc[df["distance"].between(10000,20000,inclusive="left"), "distancia_centro"] = "Entre 10 y 20 km"
df.loc[df["distance"].between(20000,30000,inclusive="left"), "distancia_centro"] = "Entre 20 y 30 km"
df.loc[df["distance"].between(30000,40000,inclusive="left"), "distancia_centro"] = "Entre 30 y 40 km"
df.loc[df["distance"].between(40000,50000,inclusive="both"), "distancia_centro"] = "Entre 40 y 50 km"
df.loc[df["distance"] > 50000, "distancia_centro"] = "Mas de 50 km"

df["distancia_centro"].value_counts()

# Dropeamos distance

df.drop(columns="distance",inplace=True)

# Gestionar nulos variables categóricas
- Las booleanas las volvemos categóricas
- Los nans los pondremos en desconocido

df.info()

### propertyType

df["propertyType"].isnull().sum()

### status

df["status"].isnull().sum()

df["status"]= df["status"].fillna("desconocido")
df["status"].unique()

df["status"].isnull().sum()

### floor

df["floor"].isnull().sum()

df["floor"]= df["floor"].fillna("desconocido")
df["floor"].unique()

df["floor"].isnull().sum()

### neighborhood (Dropear Columna)
- Demasiados nulos, no tiene sentido

df["neighborhood"].isnull().sum()

df.drop(columns="neighborhood",inplace=True)

### district
- Tal vez especifique demasiado

df["district"].isnull().sum()

df["district"]= df["district"].fillna("desconocido")
df["district"].isnull().sum()

### province

df["province"].isnull().sum()

### municipality

df["municipality"].isnull().sum()

###  exterior
- Cambiamos de booleano a categorica

df["exterior"].isnull().sum()

df["exterior"].unique()

df.loc[df["exterior"] == True, "exterior"] = "vista exterior"
df.loc[df["exterior"] == False, "exterior"] = "vista interior"
df["exterior"].unique()

### hasLift 
- Cambiamos de booleano a categorica

df["hasLift"].isnull().sum()

df["hasLift"].unique()

df.loc[df["hasLift"] == True, "hasLift"] = "tiene ascensor"
df.loc[df["hasLift"] == False, "hasLift"] = "no tiene ascensor"
df["hasLift"] = df["hasLift"].fillna("desconocido")
df["hasLift"].unique()

df["hasLift"].isnull().sum()

### hasPlan
- Cambiamos de booleano a categorica


df["hasPlan"].isnull().sum()

df.loc[df["hasPlan"] == True, "hasPlan"] = "tiene planos"
df.loc[df["hasPlan"] == False, "hasPlan"] = "no tiene planos"
df["hasPlan"] = df["hasPlan"].fillna("desconocido")
df["hasPlan"].unique()

df["hasPlan"].isnull().sum()

### has3DTour
- Cambiamos de booleano a categorica


df["has3DTour"].isnull().sum()

df.loc[df["has3DTour"] == True, "has3DTour"] = "tiene tour 3D"
df.loc[df["has3DTour"] == False, "has3DTour"] = "no tiene tour 3D"
df["has3DTour"] = df["has3DTour"].fillna("desconocido")
df["has3DTour"].unique()

df["has3DTour"].isnull().sum()

### has360
- Cambiamos de booleano a categorica

df["has360"].isnull().sum()

c = df["has360"].isnull() == True
df[c]


#### Habrá que ver si estos datos merece la pena tenerlos, hay muchas cosas desconocidas

df.loc[df["has360"] == True, "has360"] = "tiene fotos 360"
df.loc[df["has360"] == False, "has360"] = "no tiene fotos 360"
df["has360"] = df["has360"].fillna("desconocido")
df["has360"].unique()

df["has360"].isnull().sum()

### distancia_centro

df["distancia_centro"].isnull().sum()

# Gestión nulos numéricas
- Sin tocar la VR (Precio).
- Rellenaremos con IterativeImputer y RandomForest
    - Las relaciones son complejas

df.isnull().sum()

# Modelo 7 eliminar province

df.drop(columns="province",inplace=True)

# Modelo 8 Eliminar district y has3DTour

df.drop(columns=["district","has3DTour"],inplace=True)

# Modelo 9 Eliminar status

df["status"].value_counts()

df.drop(columns="status",inplace=True)

# Modelo 11 Eliminar distancia_centro


df["distancia_centro"].value_counts()

df.drop(columns="distancia_centro",inplace=True)

# Guardamos

df.to_csv("../../datos/03_api_rent_sin_nulos.csv",index=False)
